feat: [BSZ-MODEL-01-model-service-surfaces-005] - [Preserve direct, inert Models root construction]
